http_capture/requirements.txt
fastapi
pydantic
python-dotenv
asyncio
azure-servicebus
